Add files via upload
This powerpoint presentation is an overall summary of the project with detailed description for background, literature review, data collection, methodology and results. It also demonstrates the outline of the written manuscript, though certain changes in the "Methodology and Results" part have been made. Please refer to the manuscript for final version of the project.